+++ how-to/customize-workspace/client/src/framework/shapes/interopbroker-shapes.ts
@@ -24,15 +24,9 @@ export interface IntentPickerResponse {
 export interface IntentOptions {
 	/** How long should the broker wait after launching a view/window for it to register an intent handler. The default is 5000 (5 seconds) */
 	intentTimeout?: number;
-	/** If an unregistered app is included here then it indicates you wish to support selecting views/windows that are not linked to an app from
-	 * an intent picker that supports instances. The intents and contexts in this app specify which you support for unregistered instances.
-	 * Do not specify a manifest or manifestType for this entry (we don't launch unregistered apps so there is no need for a manifest as it
-	 * is existing instances and we specify a custom manifestType for consideration when we are bringing an unregistered web app instance into focus).
-	 * */
-	unregisteredApp?: PlatformApp;
 }
 
-export interface IntentPickerOptions {
+export interface IntentResolverOptions {
 	/** The url of the html page that has the intent picker */
 	url: string;
 	/** the height you wish the window to be */
@@ -45,7 +39,19 @@ export interface IntentPickerOptions {
 	title?: string;
 }
 export interface PlatformInteropBrokerOptions {
-	/** Intent Picker configuration if you wish to support intents. It needs to support the functions required by the platform */
-	intentPicker?: IntentPickerOptions;
-	intentOptions: IntentOptions;
+	/** Intent Resolver configuration if you wish to support intents. It needs to support the functions required by the platform */
+	intentResolver?: IntentResolverOptions;
+	/** Options related to the way this platform supports intents */
+	intentOptions?: IntentOptions;
+	/** If an unregistered app is included here then it indicates you wish to support selecting views/windows that are not linked to an app from
+	 * an intent picker that supports instances. The intents and contexts in this app specify which you support for unregistered instances.
+	 * Do not specify a manifest or manifestType for this entry (we don't launch unregistered apps so there is no need for a manifest as it
+	 * is existing instances and we specify a custom manifestType for consideration when we are bringing an unregistered web app instance into focus).
+	 * */
+	unregisteredApp?: PlatformApp;
+}
+
+export interface ApiMetadata {
+	type: "fdc3" | "interop";
+	version?: "1.2" | "2.0" | string;
 }