+++ how-to/customize-workspace/client/webpack.config.js
@@ -137,33 +137,6 @@ const configs = [
 		}
 	},
 	{
-		entry: './client/src/modules/endpoints/fdc3-app/index.ts',
-		devtool: 'inline-source-map',
-		module: {
-			rules: [
-				{
-					test: /\.tsx?$/,
-					use: 'ts-loader',
-					exclude: /node_modules/
-				}
-			]
-		},
-		resolve: {
-			extensions: ['.tsx', '.ts', '.js']
-		},
-		externals: { fin: 'fin' },
-		output: {
-			filename: 'fdc3-app.bundle.js',
-			library: {
-				type: 'module'
-			},
-			path: path.resolve(__dirname, '..', 'public', 'js', 'modules', 'endpoints')
-		},
-		experiments: {
-			outputModule: true
-		}
-	},
-	{
 		entry: './client/src/modules/endpoints/inline-apps/index.ts',
 		devtool: 'inline-source-map',
 		module: {